+++ chart/spec/configuration/gitlab_shell_spec.rb
@@ -6,9 +6,7 @@ require 'hash_deep_merge'
 describe 'gitlab-shell configuration' do
   let(:t) { HelmTemplate.new(values) }
   let(:default_values) do
-    YAML.safe_load(%(
-      certmanager-issuer:
-        email: test@example.com
+    HelmTemplate.with_defaults(%(
       global: {}
       gitlab:
         gitlab-shell: