+++ chart/spec/configuration/global_spec.rb
@@ -5,9 +5,7 @@ require 'hash_deep_merge'
 
 describe 'global configuration' do
   let(:default_values) do
-    YAML.safe_load(%(
-      certmanager-issuer:
-        email: test@example.com
+    HelmTemplate.with_defaults(%(
       global: {}
     ))
   end