+++ chart/spec/configuration/ingress_spec.rb
@@ -34,10 +34,7 @@ describe 'GitLab Ingress configuration(s)' do
   end
 
   let(:default_values) do
-    YAML.safe_load(%(
-      certmanager-issuer:
-        email: test@example.com
-    ))
+    HelmTemplate.defaults
   end
 
   let(:ingress_names) do